requirements/dev.txt
-r common.txt
Fabric3==1.14.post1
ipdb==0.11